+++ masq_lib/Cargo.toml
@@ -12,9 +12,11 @@ workspace = "../node"
 actix = "0.7.9"
 clap = "2.33.0"
 lazy_static = "1.3.0"
+regex = "1.0.5"
 serde = "1.0.99"
 serde_derive = "1.0.99"
 serde_json = "1.0.40"
+tiny-hderive = "0.2.1"
 toml = "0.5.3"
 websocket = {version = "0.23.0", default-features = false, features = ["sync"]}
