+++ masq_lib/src/constants.rs
@@ -1,7 +1,5 @@
 // Copyright (c) 2019-2020, MASQ (https://masq.ai) and/or its affiliates. All rights reserved.
 
-pub const LOWEST_USABLE_INSECURE_PORT: u16 = 1025;
 pub const HIGHEST_RANDOM_CLANDESTINE_PORT: u16 = 9999;
-pub const HIGHEST_USABLE_PORT: u16 = 65535;
 pub const HTTP_PORT: u16 = 80;
 pub const TLS_PORT: u16 = 443;
