+++ masq_lib/src/crash_point.rs
