+++ node/src/database/config_dumper.rs
@@ -3,7 +3,7 @@
 use crate::bootstrapper::RealUser;
 use crate::config_dao::{ConfigDao, ConfigDaoReal};
 use crate::database::db_initializer::{DbInitializer, DbInitializerReal, DATABASE_FILE};
-use crate::node_configurator::{app_head, chain_arg, data_directory_arg, real_user_arg};
+use crate::node_configurator::{app_head};
 use crate::privilege_drop::{PrivilegeDropper, PrivilegeDropperReal};
 use clap::Arg;
 use heck::MixedCase;
@@ -12,6 +12,7 @@ use masq_lib::multi_config::{CommandLineVcl, EnvironmentVcl, MultiConfig, Virtua
 use serde_json::json;
 use serde_json::{Map, Value};
 use std::path::PathBuf;
+use masq_lib::shared_schema::{chain_arg, data_directory_arg, real_user_arg};
 
 const DUMP_CONFIG_HELP: &str =
     "Dump the configuration of MASQ Node to stdout in JSON. Used chiefly by UIs.";
