+++ node/src/node_configurator/mod.rs
@@ -7,10 +7,9 @@ pub mod node_configurator_standard;
 
 use crate::blockchain::bip32::Bip32ECKeyPair;
 use crate::blockchain::bip39::Bip39;
-use crate::blockchain::blockchain_interface::{chain_id_from_name, DEFAULT_CHAIN_NAME};
+use crate::blockchain::blockchain_interface::{chain_id_from_name};
 use crate::bootstrapper::RealUser;
 use crate::database::db_initializer::{DbInitializer, DbInitializerReal, DATABASE_FILE};
-use crate::node_configurator::node_configurator_standard::DEFAULT_UI_PORT_VALUE;
 use crate::persistent_configuration::{
     PersistentConfigError, PersistentConfiguration, PersistentConfigurationReal,
 };
@@ -31,34 +30,15 @@ use std::io::Read;
 use std::path::PathBuf;
 use std::str::FromStr;
 use tiny_hderive::bip44::DerivationPath;
+use masq_lib::shared_schema::{chain_arg, config_file_arg, data_directory_arg, real_user_arg};
 
 pub trait NodeConfigurator<T> {
     fn configure(&self, args: &Vec<String>, streams: &mut StdStreams<'_>) -> T;
 }
 
-const CHAIN_HELP: &str =
-    "The blockchain network MASQ Node will configure itself to use. You must ensure the \
-    Ethereum client specified by --blockchain-service-url communicates with the same blockchain network.";
-pub const CONFIG_FILE_HELP: &str =
-    "Optional TOML file containing configuration that doesn't often change. Should contain only \
-     scalar items, string or numeric, whose names are exactly the same as the command-line parameters \
-     they replace (except no '--' prefix). If you specify a relative path, or no path, the Node will \
-     look for your config file starting in the --data-directory. If you specify an absolute path, \
-     --data-directory will be ignored when searching for the config file. A few parameters \
-     (such as --config-file, --generate-wallet, and --recover-wallet) must not be specified in a config file.";
-pub const CONSUMING_PRIVATE_KEY_HELP: &str = "The private key for the Ethereum wallet from which you wish to pay \
-     other Nodes for routing and exit services. Mostly this is used for testing; be careful using it for real \
-     traffic, because this value is very sensitive: anyone who sees it can use it to drain your consuming wallet. \
-     If you use it, don't put it on the command line (the environment is good, the config file is less so), \
-     make sure you haven't already set up a consuming wallet with a derivation path, and make sure that you always \
-     supply exactly the same private key every time you run the Node. A consuming private key is 64 case-insensitive \
-     hexadecimal digits.";
 pub const CONSUMING_WALLET_HELP: &str = "The BIP32 derivation path for the wallet from which your Node \
      should pay other Nodes for routing and exit services. (If the path includes single quotes, enclose it in \
      double quotes.) Defaults to m/44'/60'/0'/0/0.";
-pub const DATA_DIRECTORY_HELP: &str =
-    "Directory in which the Node will store its persistent state, including at \
-     least its database and by default its configuration file as well.";
 pub const EARNING_WALLET_HELP: &str =
     "Denotes the wallet into which other Nodes will pay yours for its routing and exit services. May either be a \
      BIP32 derivation path (defaults to m/44'/60'/0'/0/1) or an Ethereum wallet address. (If the derivation path \
@@ -69,11 +49,6 @@ pub const MNEMONIC_PASSPHRASE_HELP: &str =
     "A passphrase for the mnemonic phrase. Cannot be changed later and still produce the same addresses. This is a \
      secret; providing it on the command line or in a config file is insecure and unwise. If you don't specify it anywhere, \
      you'll be prompted for it at the console.";
-pub const REAL_USER_HELP: &str =
-    "The user whose identity Node will assume when dropping privileges after bootstrapping. Since Node refuses to \
-     run with root privilege after bootstrapping, you might want to use this if you start the Node as root, or if \
-     you start the Node using pkexec or some other method that doesn't populate the SUDO_xxx variables. Use a value \
-     like <uid>:<gid>:<home directory>.";
 pub const DB_PASSWORD_HELP: &str =
     "A password or phrase to encrypt your consuming wallet in the MASQ Node database or decrypt a keystore file. Can be changed \
      later and still produce the same addresses. This is a secret; providing it on the command line or in a config file is \
@@ -94,15 +69,6 @@ pub fn app_head() -> App<'static, 'static> {
 
 // These Args are needed in more than one clap schema. To avoid code duplication, they're defined here and referred
 // to from multiple places.
-pub fn config_file_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name("config-file")
-        .long("config-file")
-        .value_name("FILE-PATH")
-        .default_value("config.toml")
-        .takes_value(true)
-        .required(false)
-        .help(CONFIG_FILE_HELP)
-}
 
 pub fn consuming_wallet_arg<'a>() -> Arg<'a, 'a> {
     Arg::with_name("consuming-wallet")
@@ -111,28 +77,6 @@ pub fn consuming_wallet_arg<'a>() -> Arg<'a, 'a> {
         .empty_values(false)
         .validator(common_validators::validate_derivation_path)
         .help(&CONSUMING_WALLET_HELP)
-}
-
-pub fn data_directory_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name("data-directory")
-        .long("data-directory")
-        .value_name("DATA-DIRECTORY")
-        .required(false)
-        .takes_value(true)
-        .empty_values(false)
-        .help(DATA_DIRECTORY_HELP)
-}
-
-pub fn chain_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name("chain")
-        .long("chain")
-        .value_name("CHAIN")
-        .min_values(1)
-        .max_values(1)
-        .takes_value(true)
-        .possible_values(&["dev", "mainnet", "ropsten"])
-        .default_value(DEFAULT_CHAIN_NAME)
-        .help(CHAIN_HELP)
 }
 
 pub fn earning_wallet_arg<F>(help: &str, validator: F) -> Arg
@@ -170,49 +114,6 @@ pub fn mnemonic_passphrase_arg<'a>() -> Arg<'a, 'a> {
         .min_values(0)
         .max_values(1)
         .help(MNEMONIC_PASSPHRASE_HELP)
-}
-
-#[cfg(not(target_os = "windows"))]
-pub fn real_user_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name("real-user")
-        .long("real-user")
-        .value_name("REAL-USER")
-        .required(false)
-        .takes_value(true)
-        .validator(common_validators::validate_real_user)
-        .help(REAL_USER_HELP)
-}
-
-#[cfg(target_os = "windows")]
-pub fn real_user_arg<'a>() -> Arg<'a, 'a> {
-    Arg::with_name("real-user")
-        .long("real-user")
-        .value_name("REAL-USER")
-        .required(false)
-        .takes_value(true)
-        .validator(common_validators::validate_real_user)
-        .hidden(true)
-}
-
-pub fn ui_port_arg(help: &str) -> Arg {
-    Arg::with_name("ui-port")
-        .long("ui-port")
-        .value_name("UI-PORT")
-        .takes_value(true)
-        .default_value(&DEFAULT_UI_PORT_VALUE)
-        .validator(crate::node_configurator::common_validators::validate_ui_port)
-        .help(help)
-}
-
-pub fn db_password_arg(help: &str) -> Arg {
-    Arg::with_name("db-password")
-        .long("db-password")
-        .value_name("DB-PASSWORD")
-        .required(false)
-        .takes_value(true)
-        .min_values(0)
-        .max_values(1)
-        .help(help)
 }
 
 pub fn determine_config_file_path(app: &App, args: &Vec<String>) -> (PathBuf, bool) {
@@ -552,7 +453,6 @@ pub fn flushed_write(target: &mut dyn io::Write, string: &str) {
 }
 
 pub mod common_validators {
-    use masq_lib::constants::LOWEST_USABLE_INSECURE_PORT;
     use regex::Regex;
     use tiny_hderive::bip44::DerivationPath;
 
@@ -779,6 +679,7 @@ mod tests {
     use std::io::Cursor;
     use std::sync::{Arc, Mutex};
     use tiny_hderive::bip44::DerivationPath;
+    use masq_lib::shared_schema::db_password_arg;
 
     #[test]
     fn validate_ethereum_address_requires_an_address_that_is_42_characters_long() {
